Re-factor tests into TestMain/Tester format files, separating Time and TimeInterval
